+++ extras/tests/ContinuousStepperTests/powerOff.cpp
@@ -6,14 +6,14 @@
 TEST_CASE("ContinuousStepper::powerOff()") {
   ContinuousStepper stepper;
 
-  GIVEN("a stepper with enable pin configured") {
+  GIVEN("begin(10, 11, 12) was called") {
     stepper.begin(10, 11, 12);
-    CLEAR_ARDUINO_LOG();
 
     WHEN("powerOff() is called") {
+      CLEAR_ARDUINO_LOG();
       stepper.powerOff();
 
-      THEN("it should set enable pin to LOW") {
+      THEN("it should set pin 12 to LOW") {
         CHECK_ARDUINO_LOG({
             {0'000, "digitalWrite(12, LOW)"},
         });
@@ -21,32 +21,49 @@ TEST_CASE("ContinuousStepper::powerOff()") {
     }
   }
 
-  GIVEN("a stepper with no enable pin") {
+  GIVEN("begin(10, 11)") {
     stepper.begin(10, 11);
-    CLEAR_ARDUINO_LOG();
 
     WHEN("powerOff() is called") {
+      CLEAR_ARDUINO_LOG();
       stepper.powerOff();
 
       THEN("it should do nothing") {
         CHECK_ARDUINO_LOG({});
       }
     }
-  }
 
-  GIVEN("a spinning stepper") {
-    stepper.begin(10, 11);
-    stepper.spin(10);
-    loop_till(stepper, 100'000);
-    REQUIRE(stepper.isSpinning() == true);
-    REQUIRE(stepper.speed() == 10);
+    AND_GIVEN("spin(10) was called") {
+      stepper.spin(10);
+      loop_till(stepper, 100'000);
+      REQUIRE(stepper.isSpinning() == true);
+      REQUIRE(stepper.speed() == 10);
 
-    WHEN("powerOff() is called") {
-      stepper.powerOff();
+      WHEN("powerOff() is called") {
+        CLEAR_ARDUINO_LOG();
+        stepper.powerOff();
 
-      THEN("it should abruptly stop the rotation") {
-        CHECK(stepper.isSpinning() == false);
-        CHECK(stepper.speed() == 0);
+        THEN("it should abruptly stop the rotation") {
+          loop_till(stepper, 100'000);
+          CHECK(stepper.isSpinning() == false);
+          CHECK(stepper.speed() == 0);
+          CHECK_ARDUINO_LOG({});
+        }
+      }
+    }
+
+    AND_GIVEN("setEnablePin(12) was called") {
+      stepper.setEnablePin(12);
+
+      WHEN("powerOff() is called") {
+        CLEAR_ARDUINO_LOG();
+        stepper.powerOff();
+
+        THEN("it should set pin 12 to LOW") {
+          CHECK_ARDUINO_LOG({
+              {0, "digitalWrite(12, LOW)"},
+          });
+        }
       }
     }
   }